+++ src/agents/embedded-agent-runner/extra-params.ts
@@ -33,8 +33,8 @@ import {
 } from "../../plugins/provider-hook-runtime.js";
 import type { ProviderRuntimeModel } from "../../plugins/provider-runtime-model.types.js";
 import { resolveModelExtraParamSources } from "../model-extra-params.js";
-import { canonicalizeMaxTokensParam, resolveMaxTokensParam } from "../model-max-tokens-params.js";
-import { detectOpenAICompletionsCompat } from "../openai-completions-compat.js";
+import { canonicalizeMaxTokensParam, resolveMaxTokensParam } from ".../../packages/ai/src/transports/model-max-tokens-params.js";
+import { detectOpenAICompletionsCompat } from ".../../packages/ai/src/transports/openai-completions-compat.js";
 import { supportsGptParallelToolCallsPayload } from "../provider-api-families.js";
 import { resolveProviderRequestPolicyConfig } from "../provider-request-config.js";
 import type { AgentRuntimeTransport } from "../runtime-plan/types.js";
