+++ src/agents/openai-completions-compat.test.ts
@@ -1,6 +1,6 @@
 // Verifies OpenAI-compatible endpoint defaults for streaming usage and reasoning payloads.
 import { describe, expect, it } from "vitest";
-import { detectOpenAICompletionsCompat } from "./openai-completions-compat.js";
+import { detectOpenAICompletionsCompat } from "../../packages/ai/src/transports/openai-completions-compat.js";
 import { resolveOpenAICompletionsCompatDefaults } from "./openai-completions-compat.test-support.js";
 
 describe("resolveOpenAICompletionsCompatDefaults", () => {
