+++ src/agents/openai-responses-payload-policy.test.ts
@@ -7,7 +7,7 @@ import { describe, expect, it } from "vitest";
 import {
   applyOpenAIResponsesPayloadPolicy,
   resolveOpenAIResponsesPayloadPolicy,
-} from "./openai-responses-payload-policy.js";
+} from "../../packages/ai/src/transports/openai-responses-payload-policy.js";
 
 describe("openai responses payload policy", () => {
   it("forces store for native OpenAI responses payloads but keeps disable mode for transport defaults", () => {
