+++ src/agents/openai-transport-params.code-mode.test.ts
@@ -2,7 +2,7 @@ import { describe, expect, it } from "vitest";
 import {
   assertCodeModeResponsesToolSurface,
   enforceCodeModeResponsesToolSurface,
-} from "./openai-transport-params.js";
+} from "../../packages/ai/src/transports/openai-transport-params.js";
 
 describe("OpenAI Code Mode direct tools", () => {
   it("keeps the native image loader model-visible", () => {
